In the error message, print value (%v) instead of assuming a string or byte slice (%s)
StatusText says it returns an string
But
http.StatusOK is returning a int
resp.StatusCode is returning a int

Since this is an error message, if it fails, might as well give the value as we received it in the error message.

"data" is already being converted, so the %s works.